requirements.txt
fastapi
uvicorn
asgiref
click
colorama
fastapi
fastapi-jwt-auth
h11
pydantic
PyJWT
starlette
typing-extensions
uvicorn
sqlalchemy
psycopg2-binary 
python-dotenv
pandas
plotly
numerize
pandas
numpy